+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog for pygments_better_html
 ==================================
 
+v0.1.1
+------
+
+* Fix table mode in Safari (with an ugly hack).
+* Accept True and False as linenos values for compatibility.
+* Accept a list of wrapper classes for `get_style_defs`.
+
 v0.1.0
 ------